Made Pectinid Plotly friendly
This is the main plotly implementation revision to pectinid.  Still having various minor issues with colors, but everything appears to be fully functional now.  The only thing that changes is there is a plot_as input, it defaults to 'ggplot' so old code should work as always without specfying anything.  Also changes to combine_shapefile_layers, know known as combo_shp both the function call the the file name, added option to force the output to be a "MULTILINESTRING".  Super minor change to conver_coords which allows a coulpe more options for names when trying to plot all the way to NFLD.  Note that at the time of this revision there was a bug in marmap which was causing issues so the bathy extent was hard coded, something we'll want to change back when bug is fixed.