Custom changes in pecjector+ new function for INLA mesh
Created (stole) a new function that converts INLA mesh into a sf object, thanks to Finn Lundgren.  Also, made some changes to allow for sf and sp object to be loaded right into pecjector via the custom call.  For now this is pretty simple, probably going to want to expand this so we can add custom colors and fills to this object, shouldn't be too hard just need an example (likely inshore contours).