Pectinid custom layer update to prepare to inshore contours
Updated Pectinid to include an add_custom layer, the behaviour of the add_custom mirrors what is done with add_inla but is also more flexible as add_custom does not require your custom layer to have any fill.  Also included a txt.size option to you can control the size of text if you so desire, and the option to plot a legend for either the INLA or custom layers if needed. Note that continous color fills are not supported with plotly option, if you try to make a continous colorf fill ramp it just makes it a discrete ramp (same with add_inla). Legends in plotly not showing up, not a biggy, but just a note. Ongoing bug fixes as well.