Pecjector expand issue resolved (kinda)
Seemed to be a combination of issues.  Using 'area' as an object in the call to convert.coords() while it was also a function (in terra package) seemed to cause a problem.  In addition, when clipping the bounding box, it seems that sf() thought there was a 'point' in the polygon and it was breaking.  That issue not fully resolved, just made our default boundaries not have this issue.